# src/main/java/com/github/mahmoudalikhalil/elm/model/entity/Product.java
package com.github.mahmoudalikhalil.elm.model.entity;

import com.github.mahmoudalikhalil.elm.constant.Constant;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.hibernate.annotations.CreationTimestamp;
import org.hibernate.annotations.DynamicUpdate;
import org.hibernate.annotations.UpdateTimestamp;

import java.io.Serializable;
import java.math.BigDecimal;
import java.time.Instant;



@Entity
@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
@DynamicUpdate
@NamedEntityGraph(name = Constant.PRODUCT_DEALER_GRAPH_NAME, attributeNodes = @NamedAttributeNode(Product_.DEALER))
public class Product implements Serializable {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String name;
    private BigDecimal price;
    @Enumerated(EnumType.STRING)
    private Status status;
    @JoinColumn(name = "user_id")
    @ManyToOne
    private User dealer;
    @Column(updatable = false)
    @CreationTimestamp
    private Instant createdAt;
    @UpdateTimestamp
    private Instant updatedAt;
}